+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.0.3 - 2017-10-19
+### Changed
+- Set `set_ignore_parser_errors` to `true` to make it more lenient at parsing invalid HTML from Redactor
+- Synced up the `config.php` and `Settings` model with the updated `php-typography` lib
+
 ## 1.0.2 - 2017-10-18
 ### Changed
 - Switched to the more modern and maintained `mundschenk-at/php-typography` package